pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "captn"
description = "Python client to interact with captn services"
# readme = "README.md"
authors = [
    { name = "airt", email = "info@airt.ai" },
]

keywords = ["captn", "autogen"]

requires-python = ">=3.11,<3.12"

classifiers = [
    "Development Status :: 5 - Production/Stable",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Operating System :: OS Independent",
    "Topic :: Internet",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development",
    "Typing :: Typed",
    "Intended Audience :: Developers",
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "Environment :: Web Environment",
    "Framework :: AsyncIO",
    "Framework :: Pydantic",
    "Framework :: Pydantic :: 1",
    "Framework :: Pydantic :: 2",
]

dynamic = ["version"]

dependencies = [
    "requests>=2.20.0,<3.0.0",
]

[project.optional-dependencies]

# dev dependencies
lint = [
    "types-PyYAML",
    "types-setuptools",
    "types-ujson",
    "types-Pygments",
    "types-docutils",
    "types-requests",
    "mypy==1.10.1",
    "black==24.4.2",
    "isort>=5",
    "ruff==0.5.0",
    "pyupgrade-directories",
    "bandit==1.7.9",
    "semgrep==1.75.0",
    "pre-commit==3.7.1",
    "detect-secrets==1.5.0",
]

test-core = [
    "coverage[toml]==7.5.4",
    "pytest==8.2.2",
    "pytest-asyncio>=0.23.6",
    "dirty-equals==0.7.1.post0",
    "pyarrow==16.1.0",
    "pytest-retry==1.6.3",
    "pytest-repeat==0.9.3",
]

testing = [
    "captn[test-core]",
    "watchfiles",
    "pytest-recording",
    "freezegun"
]

benchmarking = [
    "typer==0.9.4", # downgraded because fastagency
    "filelock==3.15.4",
    "tabulate==0.9.0",
]

agents = [
    "fastapi>=0.110.2,<0.111.0", # downgraded because fastagency
    "APScheduler==3.10.4",
    "prisma==0.13.1",
    "google-ads==24.1.0",
    "httpx==0.27.0",
    "uvicorn==0.30.1",
    "python-dotenv==1.0.1",
    "pyautogen[websurfer,websockets]==0.2.31",
    "pandas>=2.1",
    "fastcore==1.5.48",
    "asyncer==0.0.7",
    "pydantic==2.7.4",
    "markdownify==0.12.1",
    "tenacity==8.4.2",
    "prometheus-client==0.20.0",
    "opentelemetry-distro==0.46b0",
    "opentelemetry-instrumentation-fastapi==0.46b0",
    "opentelemetry-instrumentation-logging==0.46b0",
    "opentelemetry-exporter-otlp==1.25.0",
    "openpyxl==3.1.5",
    "aiofiles==24.1.0",
    "fastagency@git+https://github.com/airtai/fastagency.git@3a2346c",
    "python-multipart==0.0.9", # remove after fastagency is updated
]

dev = [
    "captn[lint,testing,agents,benchmarking]",
]

[project.scripts]
benchmark = "captn.captn_agents.backend.benchmarking.base:app"

[project.urls]
Homepage = "https://github.com/airtai/captn-backend"
Documentation = "https://github.com/airtai/captn-backend"
Tracker = "https://github.com/airtai/captn-backend/issues"
Source = "https://github.com/airtai/captn-backend"

[tool.hatch.version]
path = "captn/__about__.py"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build]
skip-excluded-dirs = true
exclude = [
    "/tests",
    "/docs",
]

[tool.mypy]
strict = true
python_version = "3.11"
ignore_missing_imports = true
install_types = true
non_interactive = true
plugins = ["pydantic.mypy"]

files = [
    "captn",
    "google_ads",
    "openai_agent",
    "tests",
    "application.py",
]

# from https://blog.wolt.com/engineering/2021/09/30/professional-grade-mypy-configuration/
disallow_untyped_defs = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
show_error_codes = true
warn_unused_ignores = true

disallow_incomplete_defs = true
disallow_untyped_decorators = true
disallow_any_unimported = false

[tool.isort]
profile = "black"
known_first_party = ["captn"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.ruff]
fix = true
line-length = 88
target-version = "py311"
include = ["captn/**/*.py", "google_ads/**/*.py", "openai_agent/**/*.py", "tests/**/*.py", "application.py/**/*.py", "pyproject.toml"]
exclude = []

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "C",  # flake8-comprehensions
    "B",  # flake8-bugbear
    "Q",  # flake8-quotes
]
ignore = [
    "E501",  # line too long, handled by ruff format
    "C901",  # too complex
]

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = ["fastapi.Depends"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.format]

docstring-code-format = true

[tool.pytest.ini_options]

minversion = "7.0"

addopts = "-vv"

testpaths = [
    "tests/ci",
]

markers = [
    "openai: mark test as using openai.",
    "brief_creation_team: maker test as end2end testing of the brief_creation_team",
    "campaign_creation_team: maker test as end2end testing of the campaign_creation_team",
    "google_ads_team: maker test as end2end testing of the google_ads_team",
    "weekly_analysis_team: maker test as end2end testing of the weekly_analysis_team",
    "get_info_from_the_web_page: maker test as end2end testing of the websurfer",
]

# needed for pytest-retry
retry_delay = 1.0
cumulative_timing = false

[tool.coverage.run]
parallel = true
branch = true
concurrency = [
    "multiprocessing",
    "thread"
]
source = [
    "captn",
    # "tests"
]
context = '${CONTEXT}'
omit = [
    "**/__init__.py",
    "tests/mypy/*",
]

[tool.coverage.report]
show_missing = true
skip_empty = true
exclude_lines = [
    ".*# pragma: no cover",
    "if __name__ == .__main__.:",
    "self.logger",
    "def __repr__",
    "lambda: None",
    "from .*",
    "import .*",
    '@(abc\.)?abstractmethod',
    "raise NotImplementedError",
    'raise AssertionError',
    'raise ValueError',
    'logger\..*',
    "pass",
    '\.\.\.',
]
omit = [
    '*/__about__.py',
    '*/__main__.py',
    '*/__init__.py',
    '*/annotations.py',
]

[tool.bandit]
